+++ src/templates/artifacts/SimpleJpaChildController.groovy
@@ -18,17 +18,17 @@ class $className {
     def listAll = {
 <%
     fields.each { field ->
-        if (field.info=="DOMAIN_CLASS" && !field.annotations?.containsAttribute('mappedBy')) {
+        if (isManyToOne(field) && !isOwned(field)) {
             out << "\t\t\texecInsideUIAsync {model.${field.name}List.clear() }\n"
         }
     }
 
     fields.each { field ->
-        if (field.info=="DOMAIN_CLASS" && field.annotations?.containsAttribute('mappedBy')) return
+        if (isOwned(field)) return
 
-        if (field.info=="DOMAIN_CLASS") {
+        if (isManyToOne(field)) {
             out << "\t\tList ${field.name}Result = findAll${field.type}()\n"
-        } else if (field.type.toString()=="List" && field.info!="UNKNOWN" && field.annotations?.containsAttribute('mappedBy')) {
+        } else if (isManyToMany(field)) {
             out << "\t\tList ${field.name}Result = findAll${field.info}()\n"
         } else if (field.info=="UNKNOWN") {
             out << "\t\t// ${field.name} isn't supported! It must be coded manually!\n"
@@ -36,11 +36,11 @@ class $className {
     }
 
     fields.each { field ->
-        if (field.info=="DOMAIN_CLASS" && field.annotations?.containsAttribute('mappedBy')) return
+        if (isOwned(field)) return
 
-        if (field.info=="DOMAIN_CLASS") {
+        if (isManyToOne(field)) {
             out << "\t\t\texecInsideUIAsync{ model.${field.name}List.addAll(${field.name}Result) }\n"
-        } else if (field.type.toString()=="List" && field.info!="UNKNOWN" && field.annotations?.containsAttribute('mappedBy') && field.annotations?.get("OneToMany")==null) {
+        } else if (isManyToMany(field)) {
             out << "\t\t\texecInsideUIAsync{ model.${field.name}.replaceValues(${field.name}Result) }\n"
         }
     }
@@ -49,11 +49,13 @@ class $className {
     def save = {
         ${domainClass} ${domainClassAsProp} = new ${domainClass}(<%
     out << fields.collect { field ->
-        if (field.info=="DOMAIN_CLASS" && field.annotations?.containsAttribute('mappedBy')) {
+        if (isOwned(field)) {
             return "'${field.name}': null"
-        } else if (field.info=="DOMAIN_CLASS") {
+        } else if (isManyToOne(field)) {
             return "'${field.name}': model.${field.name}.selectedItem"
-        } else if (field.type.toString()=="List" && field.info!="UNKNOWN" && field.annotations?.get("OneToMany")==null) {
+        } else if (isOneToMany(field)) {
+            return "'${field.name}': new ArrayList(model.${field.name})"
+        } else if (isManyToMany(field)) {
             return "'${field.name}': model.${field.name}.selectedValues"
         } else {
             return "'${field.name}': model.${field.name}"
@@ -70,11 +72,14 @@ class $className {
             ${domainClass} selected${domainClass} = model.${domainClassAsProp}Selection.selected[0]
 <%
     fields.each { field ->
-        if (field.info=="DOMAIN_CLASS" && field.annotations?.containsAttribute('mappedBy')) return ''
+        if (isOwned(field)) return
 
-        if (field.info=="DOMAIN_CLASS") {
+        if (isManyToOne(field)) {
             out << "\t\t\tselected${domainClass}.${field.name} = model.${field.name}.selectedItem\n"
-        } else if (field.type.toString()=="List" && field.info!="UNKNOWN" && field.annotations?.get("OneToMany")==null) {
+        } else if (isOneToMany(field)) {
+            out << "\t\t\tselected${domainClass}.${field.name}.clear()\n"
+            out << "\t\t\tselected${domainClass}.${field.name}.addAll(model.${field.name})\n"
+        } else if (isManyToMany(field)) {
             out << "\t\t\tselected${domainClass}.${field.name}.clear()\n"
             out << "\t\t\tselected${domainClass}.${field.name}.addAll(model.${field.name}.selectedValues)\n"
         } else {