+++ release_notes.md
@@ -38,5 +38,6 @@ First release into griffon artifacts portal.
 7.  Add new functions that can be used inside template renderer: numberFormat, percentFormat, currentFormat, titleCase.
 8.  Add new softDelete(entity) method.
 9.  Add columnClasses to eventTableModel() to specify class type for each column (used by renderers).
-10. generate-all now support one-to-many relationship.
-11. generate-all will not generate mappedBy attributes (owned).
+10.  generate-all now support one-to-many relationship.
+11.  generate-all will not generate mappedBy attributes (owned).
+12.  generate-all will generate one-to-one relationship as pair dialog (similiar to one-to-many child dialog).